requirements.txt
streamlit
google-generativeai
PyPDF2
sentence-transformers
faiss-cpu
numpy
nltk
tqdm
